+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## UNRELEASED
 
-- Added support for budoux v0.6.2.
+- Added support Simplified Chinese
+- Added support Traditional Chinese
+- Added support Thai
+- Added support for budoux v0.6.2
 - Improved Ja model <https://github.com/google/budoux/pull/410>
 
 ## 1.1.0